requirements.txt
playwright
python-dotenv
openai